Computes a transform using the 'iplogicle' function

Transforms FCS data using the iplogicle function from FCSTrans by Quian et al. The core functionality of FCSTrans has been imported to produce transformed FCS data rescaled and truncated as produced by FCSTrans. The w parameter is estimated by iplogicle automatically, then makes a call to iplogicore which in turn uses the logicle transform code of Wayne Moore.

FCSTransTransform(transformationId = "defaultFCSTransTransform",
                  channelrange = 2^18, channeldecade = 4.5,
                  range = 4096, cutoff = -111, w = NULL, rescale = TRUE)

Arguments

transformationId
A name to assign to the transformation. Used by the transform/filter routines.

channelrange
is the range of the data. By default, 2^18 = 262144.

channeldecade
is the number of logarithmic decades. By default, it is set to 4.5.

range
the target resolution. The default value is 2^12 = 4096.

cutoff
a threshold below which the logicle transformation maps values to 0.

w
the logicle width. This is estimated by iplogicle by default. Details can be found in the Supplementary File from Quian et al.

rescale
logical parameter whether or not the data should be rescaled to the number of channels specified in range. By default, the value is TRUE but can be set to FALSE if you want to work on the transformed scale.

Details

For the details of the FCSTrans transformation, we recommend the excellent Supplementary File that accompanies Quian et al. (2012): http://onlinelibrary.wiley.com/doi/10.1002/cyto.a.22037/suppinfo

References

Y Quian, Y Liu, J Campbell, E Thompson, YM Kong, RH Scheuermann; FCSTrans: An open source software system for FCS file conversion and data transformation. Cytometry A, 2012

See also

Examples

data(GvHD)
samp <- GvHD[[1]]
## User defined logicle function
lgcl <- transformList(c('FL1-H', 'FL2-H'), FCSTransTransform())
after <- transform(samp, lgcl)

#> [1] "params: r= 262144 d= 4.5 range= 4096 cutoff= -111 w= 1.16434647887945"
#> [1] "params: r= 262144 d= 4.5 range= 4096 cutoff= -111 w= 1.16434647887945"
#> [1] "params: r= 262144 d= 4.5 range= 4096 cutoff= -111 w= 1.16434647887945"
#> [1] "params: r= 262144 d= 4.5 range= 4096 cutoff= -111 w= 1.16434647887945"