Class "compensatedParameter"

Emission spectral overlap can be corrected by subtracting the amount of spectral overlap from the total detected signals. This compensation process can be described by using spillover matrices.

Details

The compensatedParameter class allows for compensation of specific parameters the user is interested in by creating compensatedParameter objects and evaluating them. This allows for use of compensatedParameter in gate definitions.

Slots

.Data

Object of class "function".

parameters

Object of class "character" -- the flow parameters to be compensated.

spillRefId

Object of class "character" -- the name of the compensation object (The compensation object contains the spillover Matrix).

searchEnv

Object of class "environment" -environment in which the compensation object is defined.

transformationId

Object of class "character" -- a unique Id to reference the compensatedParameter object.

Note

The transformation object can be evaluated using the eval method by passing the data frame as an argument. The transformed parameters are returned as a matrix with a single column. (See example below)

Objects from the Class

Objects can be created by calls to the constructor of the form compensatedParameter(parameters,spillRefId,transformationId,searchEnv).

Extends

Class "transform", directly. Class "transformation", by class "transform", distance 2. Class "characterOrTransformation", by class "transform", distance 3.

See also

Examples

samp   <- read.flowSet(path=system.file("extdata", "compdata", "data", package="flowCore"))
cfile <- system.file("extdata","compdata","compmatrix", package="flowCore")
comp.mat <- read.table(cfile, header=TRUE, skip=2, check.names = FALSE)
comp.mat

#>      FL1-H    FL2-H   FL3-H   FL4-H
#> 1 1.000000 0.240000 0.03200 0.00113
#> 2 0.007770 1.000000 0.14000 0.00274
#> 3 0.008690 0.170000 1.00000 0.21000
#> 4 0.000795 0.000995 0.00323 1.00000

## create a compensation object 
comp <- compensation(comp.mat,compensationId="comp1")
## create a compensated parameter object 
cPar1<-compensatedParameter(c("FL1-H","FL3-H"),"comp",searchEnv=.GlobalEnv)
compOut<-eval(cPar1)(exprs(samp[[1]]))

#> Error in eval(cPar1)(exprs(samp[[1]])): trying to get slot "spillover" from an object of a basic class ("NULL") with no slots