Methods to apply functions over flowFrame margins

Returns a vector or array of values obtained by applying a function to the margins of a flowFrame. This is equivalent of running apply on the output of exprs(flowFrame).

each_col(x, FUN, ...)
each_row(x, FUN, ...)

Arguments

x
Object of class flowFrame.

FUN
the function to be applied. In the case of functions like '+', '%*%', etc., the function name must be backquoted or quoted.

...
optional arguments to 'FUN'.

See also

Examples

samp <- read.FCS(system.file("extdata", "0877408774.B08", package="flowCore"),
transformation="linearize")
each_col(samp, summary)

#>             FSC-H     SSC-H      FL1-H      FL2-H       FL3-H     FL1-A
#> Min.      85.0000   11.0000    1.00000    1.00000    1.000000    0.0000
#> 1st Qu.  385.0000  141.0000    8.13123   12.07901    2.246790    0.0000
#> Median   441.0000  189.0000  135.16485   22.46790    5.674221   26.0000
#> Mean     491.9644  277.9105  157.79417  105.98637    8.464880   34.0766
#> 3rd Qu.  518.0000  270.0000  241.44182   50.93675   10.746078   51.0000
#> Max.    1023.0000 1023.0000 3651.74127 9910.45856 3278.121151 1023.0000
#>              FL4-H   Time
#> Min.       1.00000   1.00
#> 1st Qu.    6.61169 122.00
#> Median    12.29826 288.00
#> Mean     140.39784 294.77
#> 3rd Qu.   33.37625 457.50
#> Max.    9821.71889 626.00