Class filterReference

A reference to another filter inside a reference. Users should generally not be aware that they are using this class.

Slots

name

The R name of the referenced filter.

env

The environment where the filter must live.

filterId

The filterId, not really used since you always resolve.

Objects from the Class

Objects are generally not created by users so there is no constructor function.

Extends

Class "filter", directly.