Class "filterSummary"

Class and methods to handle the summary information of a gating operation.

# S4 method for filterResult
summary(object, ...)

Arguments

object
An object inheriting from class filterResult which is to be summarized.

...
Further arguments that are passed to the generic.

Value

An object of class filterSummary for the summary constructor, a named list for the subsetting operators. The $ operator returns a named vector of the respective value, where each named element corresponds to one sub-population.

Details

Calling summary on a filterResult object prints summary information on the screen, but also creates objects of class filterSummary for computational access.

Slots

name

Object of class "character" The name(s) of the populations created in the filtering operation. For a logicalFilterResult this is just a single value; the name of the link{filter}.

true

Object of class "numeric". The number of events within the population(s).

count

Object of class "numeric". The total number of events in the gated flowFrame.

p

Object of class "numeric" The percentage of cells in the population(s).

Objects from the Class

Objects are created by calling summary on a link{filterResult} object. The user doesn't have to deal with manual object instantiation.

Methods

[[

signature(x = "filterSummary", i = "numeric"): Subset the filterSummary to a single population. This only makes sense for multipleFilterResults. The output is a list of summary statistics.

[[

signature(x = "filterSummary", i = "character"): see above

$

signature(x = "filterSummary", name = "ANY"): A list-like accessor to the slots and more. Valid values are n and count (those are identical), true and in (identical), false and out (identical), name, p and q (1-p).

coerce

signature(from = "filterSummary", to = "data.frame"): Coerce object to data.frame.

length

signature(x = "filterSummary"): The number of populations in the fitlerSummary.

names

signature(x = "filterSummary"): The names of the populations in the filterSummary.

print

signature(x = "filterSummary"): Print details about the object.

show

signature(object = "filterSummary"): Print details about the object.

toTable

signature(x = "filterSummary"): Coerce object to data.frame.

See also

Examples

library(flowStats)

## Loading example data, creating and applying a curv1Filter
dat <- read.FCS(system.file("extdata","0877408774.B08",
package="flowCore"))
c1f <- curv1Filter(filterId="myCurv1Filter", x=list("FSC-H"), bwFac=2)
fres <- filter(dat, c1f)

## creating and showing the summary
summary(fres)

#> rest: 3581 of 10000 events (35.81%)
#> peak 1: 331 of 10000 events (3.31%)
#> peak 2: 5575 of 10000 events (55.75%)
#> peak 3: 455 of 10000 events (4.55%)
#> peak 4: 58 of 10000 events (0.58%)

s <- summary(fres)

## subsetting
s[[1]]

#> $name
#> [1] "rest"
#> 
#> $true
#> [1] 3581
#> 
#> $false
#> [1] 6419
#> 
#> $count
#> [1] 10000
#> 
#> $p
#> [1] 0.3581
#> 
#> $q
#> [1] 0.6419
#> 

s[["peak 2"]]

#> $name
#> [1] "peak 2"
#> 
#> $true
#> [1] 5575
#> 
#> $false
#> [1] NA
#> 
#> $count
#> [1] 10000
#> 
#> $p
#> [1] 0.5575
#> 
#> $q
#> [1] 0.4425
#> 

##accessing details
s$true

#>   rest peak 1 peak 2 peak 3 peak 4 
#>   3581    331   5575    455     58 

s$n

#> [1] 10000

toTable(s)

#>        percent count true false      p      q
#> rest     35.81 10000 3581  6419 0.3581 0.6419
#> peak 1    3.31 10000  331  9669 0.0331 0.9669
#> peak 2   55.75 10000 5575  4425 0.5575 0.4425
#> peak 3    4.55 10000  455  9545 0.0455 0.9545
#> peak 4    0.58 10000   58  9942 0.0058 0.9942