Class "filterSummaryList"

Class and methods to handle summary statistics for from filtering operations on whole flowSets.

Arguments

object
An object of class. filterResultList which is to be summarized.

...
Further arguments that are passed to the generic.

Value

An object of class filterSummaryList.

Details

Calling summary on a filterResultList object prints summary information on the screen, but also creates objects of class filterSummaryList for computational access.

Slots

.Data

Object of class "list". The class directly extends list, and this slot holds the list data.

Usage

summary(object, ...)

Objects from the Class

Objects are created by calling summary on a link{filterResultList} object. The user doesn't have to deal with manual object instantiation.

Extends

Class "list", from .Data part.

Methods

toTable

signature(x = "filterSummaryList"): Coerce object to data.frame. Additional factors are added to indicate list items in the original object.

See also

Examples

library(flowStats)

## Loading example data, creating and applying a curv1Filter
data(GvHD)
dat <- GvHD[1:3]
c1f <- curv1Filter(filterId="myCurv1Filter", x=list("FSC-H"), bwFac=2)
fres <- filter(dat, c1f)

## creating and showing the summary
summary(fres)

#> filter summary for frame 's5a01'
#>  rest: 1002 of 3420 events (29.30%)
#>  peak 1: 2418 of 3420 events (70.70%)
#> 
#> filter summary for frame 's5a02'
#>  rest: 1026 of 3405 events (30.13%)
#>  peak 1: 2379 of 3405 events (69.87%)
#> 
#> filter summary for frame 's5a03'
#>  rest: 369 of 3435 events (10.74%)
#>  peak 1: 3066 of 3435 events (89.26%)
#> 

s <- summary(fres)

#> filter summary for frame 's5a01'
#>  rest: 1002 of 3420 events (29.30%)
#>  peak 1: 2418 of 3420 events (70.70%)
#> 
#> filter summary for frame 's5a02'
#>  rest: 1026 of 3405 events (30.13%)
#>  peak 1: 2379 of 3405 events (69.87%)
#> 
#> filter summary for frame 's5a03'
#>  rest: 369 of 3435 events (10.74%)
#>  peak 1: 3066 of 3435 events (89.26%)
#> 

## subsetting
s[[1]]

#> rest: 1002 of 3420 events (29.30%)
#> peak 1: 2418 of 3420 events (70.70%)

##accessing details
toTable(s)

#>   sample population  percent count true false         p         q
#> 1  s5a01       rest 29.29825  3420 1002  2418 0.2929825 0.7070175
#> 2  s5a01     peak 1 70.70175  3420 2418  1002 0.7070175 0.2929825
#> 3  s5a02       rest 30.13216  3405 1026  2379 0.3013216 0.6986784
#> 4  s5a02     peak 1 69.86784  3405 2379  1026 0.6986784 0.3013216
#> 5  s5a03       rest 10.74236  3435  369  3066 0.1074236 0.8925764
#> 6  s5a03     peak 1 89.25764  3435 3066   369 0.8925764 0.1074236