Class hyperlogtGml2

Hyperlog transformation parameterized according to Gating-ML 2.0.

Details

hyperlogtGml2 is defined by the following function: $$bound(hyperlog, boundMin, boundMax) = max(min(hyperlog,boundMax),boundMin))$$ where $$hyperlog(x, T, W, M, A) = root(EH(y, T, W, M, A) - x)$$ and \(EH\) is defined as: $$EH(y, T, W, M, A) = ae^{by} + cy - f$$ where

x is the value that is being transformed (an FCS dimension value). Typically, x is less than or equal to T, although the transformation function is also defined for x greater than T.

y is the result of the transformation.

T is greater than zero and represents the top of scale value.

M is greater than zero and represents the number of decades that the true logarithmic scale approached at the high end of the Hyperlog scale would cover in the plot range.

W is positive and not greater than half of M and represents the number of such decades in the approximately linear region.

A is the number of additional decades of negative data values to be included. A shall be greater than or equal to \(-W\), and less than or equal to \(M - 2W\)

root is a standard root finding algorithm (e.g., Newton's method) that finds y such as \(B(y, T, W, M, A) = x\).

and \(a\), \(b\), \(c\) and \(f\) are defined by means of \(T\), \(W\), \(M\), \(A\), \(w\), \(x0\), \(x1\), \(x2\), \(e0\), \(ca\) and \(fa\) as: $$w = W/(M+A)$$ $$x2 = A/(M+A)$$ $$x1 = x2 + w$$ $$x0 = x2 + 2*w$$ $$b = (M + A)*ln(10)$$ $$e0 = e^{b*x0}$$ $$ca= e0/w$$ $$fa = e^{b*x1} + ca*x1$$ $$a = T / (e^b + ca - fa)$$ $$c = ca * a$$ $$f = fa * a$$

In addition, if a boundary is defined by the boundMin and/or boundMax parameters, then the result of this transformation is restricted to the [boundMin,boundMax] interval. Specifically, should the result of the hyperlog function be less than boundMin, then let the result of this transformation be boundMin. Analogically, should the result of the hyperlog function be more than boundMax, then let the result of this transformation be boundMax. The boundMin parameter shall not be greater than the boundMax parameter.

Slots

.Data

Object of class function.

T

Object of class numeric -- positive constant (top of scale value).

M

Object of class numeric -- positive constant (desired number of decades).

W

Object of class numeric -- positive constant that is not greater than half of M (the number of such decades in the approximately linear region)

A

Object of class numeric -- a constant that is greater than or equal to -W, and also less than or equal to M-2W. (A represents the number of additional decades of negative data values to be included.)

parameters

Object of class "transformation" -- flow parameter to be transformed.

transformationId

Object of class "character" -- unique ID to reference the transformation.

boundMin

Object of class numeric -- lower bound of the transformation, default -Inf.

boundMax

Object of class numeric -- upper bound of the transformation, default Inf.

Note

That hyperlogtGml2 transformation brings "reasonable" data values to the scale of \([0,1]\). The transformation is somewhat similar to logicletGml2. (See Gating-ML 2.0 for detailed comparison)

The hyperlog transformation object can be evaluated using the eval method by passing the data frame as an argument. The transformed parameters are returned as a matrix with a single column. (See example below)

Objects from the Class

Objects can be created by calls to the constructor

hyperlogtGml2(parameter, T, M, W, A, transformationId, boundMin, boundMax)

Extends

Class singleParameterTransform, directly.

Class transform, by class singleParameterTransform, distance 2.

Class transformation, by class singleParameterTransform, distance 3.

Class characterOrTransformation, by class singleParameterTransform, distance 4.

References

Gating-ML 2.0: International Society for Advancement of Cytometry (ISAC) standard for representing gating descriptions in flow cytometry. http://flowcyt.sourceforge.net/gating/20141009.pdf

See also

Examples

myDataIn  <- read.FCS(system.file("extdata", "0877408774.B08",
    package="flowCore"))
myHyperLg <- hyperlogtGml2(parameters = "FSC-H", T = 1023, M = 4.5,
    W = 0.5, A = 0, transformationId="myHyperLg")
transOut  <- eval(myHyperLg)(exprs(myDataIn))