Methods to retrieve keywords of a flowFrame

Accessor and replacement methods for items in the description slot (usually read in from a FCS file header). It lists the keywords and its values for a flowFrame specified by a character vector. Additional methods for function and lists exists for more programmatic access to the keywords.

keyword(object, keyword, ...)

Arguments

object
Object of class flowFrame.

keyword
Character vector or list of potential keywords or function. If missing all keywords are returned.

...
compact: logical scaler to indicate whether to hide all the cytometer instrument and laser settings from keywords.

Details

The keyword methods allow access to the keywords stored in the FCS files, either for a flowFrame or for a list of frames in a flowSet. The most simple use case is to provide a character vector or a list of character strings of keyword names. A more sophisticated version is to provide a function which has to take one mandatory argument, the value of this is the flowFrame. This can be used to query arbitrary information from the flowFrames description slot or even the raw data. The function has to return a single character string. The list methods allow to combine functional and direct keyword access. The replacement method takes a named character vector or a named list as input. R's usual recycling rules apply when replacing keywords for a whole flowSet

Methods

keyword(object = "flowFrame", keyword = "character")

Return values for all keywords from the description slot in object that match the character vector keyword.

keyword(object = "flowFrame", keyword = "function")

Apply the function in keyword on the flowFrame object. The function needs to be able to cope with a single argument and it needs to return a single character string. A typical use case is for instance to paste together values from several different keywords or to compute some statistic on the flowFrame and combine it with one or several other keywords.

keyword(object = "flowFrame", keyword = "list")

Combine characters and functions in a list to select keyword values.

keyword(object = "flowFrame", keyword = "missing")

This is essentially an alias for description and returns all keyword-value pairs.

keyword(object = "flowSet", keyword = "list")

This is a wrapper around fsApply(object, keyword, keyword) which essentially iterates over the frames in the flowSet.

keyword(object = "flowSet", keyword = "ANY")

This first coerces the keyword (mostly a character vector) to a list and then calls the next applicable method.

See also

Examples

samp <- read.FCS(system.file("extdata","0877408774.B08", package="flowCore"))
keyword(samp)

#> $FCSversion
#> [1] "2"
#> 
#> $`$BYTEORD`
#> [1] "4,3,2,1"
#> 
#> $`$DATATYPE`
#> [1] "F"
#> 
#> $`$NEXTDATA`
#> [1] "0"
#> 
#> $`$SYS`
#> [1] "Macintosh System Software 9.2.2"
#> 
#> $CREATOR
#> [1] "CELLQuest<aa> 3.3"
#> 
#> $`$TOT`
#> [1] "10000"
#> 
#> $`$MODE`
#> [1] "L"
#> 
#> $`$PAR`
#> [1] "8"
#> 
#> $`$P1N`
#> [1] "FSC-H"
#> 
#> $`$P1R`
#> [1] "1024"
#> 
#> $`$P1B`
#> [1] "16"
#> 
#> $`$P1E`
#> [1] "0,0"
#> 
#> $`$P2N`
#> [1] "SSC-H"
#> 
#> $`$P2R`
#> [1] "1024"
#> 
#> $`$P2B`
#> [1] "16"
#> 
#> $`$P2E`
#> [1] "0,0"
#> 
#> $`$P3N`
#> [1] "FL1-H"
#> 
#> $`$P3R`
#> [1] "1024"
#> 
#> $`$P3B`
#> [1] "16"
#> 
#> $`$P3E`
#> [1] "0,0"
#> 
#> $`$P4N`
#> [1] "FL2-H"
#> 
#> $`$P4R`
#> [1] "1024"
#> 
#> $`$P4B`
#> [1] "16"
#> 
#> $`$P4E`
#> [1] "0,0"
#> 
#> $`$P5N`
#> [1] "FL3-H"
#> 
#> $`$P5R`
#> [1] "1024"
#> 
#> $`$P5B`
#> [1] "16"
#> 
#> $`$P5E`
#> [1] "0,0"
#> 
#> $`$P1S`
#> [1] "FSC-H"
#> 
#> $`$P2S`
#> [1] "SSC-H"
#> 
#> $`$P3S`
#> character(0)
#> 
#> $`$P4S`
#> character(0)
#> 
#> $`$P6N`
#> [1] "FL1-A"
#> 
#> $`$P6R`
#> [1] "1024"
#> 
#> $`$P6B`
#> [1] "16"
#> 
#> $`$P6E`
#> [1] "0,0"
#> 
#> $TIMETICKS
#> [1] "50"
#> 
#> $`$P7N`
#> [1] "FL4-H"
#> 
#> $`$P7R`
#> [1] "1024"
#> 
#> $`$P7E`
#> [1] "0,0"
#> 
#> $`$P7B`
#> [1] "16"
#> 
#> $`$P8N`
#> [1] "Time"
#> 
#> $`$P8R`
#> [1] "1024"
#> 
#> $`$P8E`
#> [1] "0,0"
#> 
#> $`$P8B`
#> [1] "16"
#> 
#> $`$P8S`
#> [1] "Time (51.20 sec.)"
#> 
#> $`SAMPLE ID`
#> [1] "Default Patient ID"
#> 
#> $`$SRC`
#> [1] "Default Patient Name"
#> 
#> $`CASE NUMBER`
#> [1] "Default Case Number"
#> 
#> $`$CYT`
#> [1] "FACSCalibur"
#> 
#> $CYTNUM
#> [1] "E5451"
#> 
#> $`$BTIM`
#> [1] "13:13:54"
#> 
#> $`$ETIM`
#> [1] "13:14:27"
#> 
#> $`BD$AcqLibVersion`
#> [1] "3.1"
#> 
#> $`BD$NPAR`
#> [1] "7"
#> 
#> $`BD$P1N`
#> [1] "FSC-H"
#> 
#> $`BD$P2N`
#> [1] "SSC-H"
#> 
#> $`BD$P3N`
#> [1] "FL1-H"
#> 
#> $`BD$P4N`
#> [1] "FL2-H"
#> 
#> $`BD$P5N`
#> [1] "FL3-H"
#> 
#> $`BD$P6N`
#> [1] "FL1-A"
#> 
#> $`BD$P7N`
#> [1] "FL4-H"
#> 
#> $`BD$WORD0`
#> [1] "24"
#> 
#> $`BD$WORD1`
#> [1] "474"
#> 
#> $`BD$WORD2`
#> [1] "594"
#> 
#> $`BD$WORD3`
#> [1] "530"
#> 
#> $`BD$WORD4`
#> [1] "700"
#> 
#> $`BD$WORD5`
#> [1] "402"
#> 
#> $`BD$WORD6`
#> [1] "401"
#> 
#> $`BD$WORD7`
#> [1] "401"
#> 
#> $`BD$WORD8`
#> [1] "402"
#> 
#> $`BD$WORD9`
#> [1] "401"
#> 
#> $`BD$WORD10`
#> [1] "299"
#> 
#> $`BD$WORD11`
#> [1] "287"
#> 
#> $`BD$WORD12`
#> [1] "557"
#> 
#> $`BD$WORD13`
#> [1] "0"
#> 
#> $`BD$WORD14`
#> [1] "483"
#> 
#> $`BD$WORD15`
#> [1] "612"
#> 
#> $`BD$WORD16`
#> [1] "546"
#> 
#> $`BD$WORD17`
#> [1] "724"
#> 
#> $`BD$WORD18`
#> [1] "104"
#> 
#> $`BD$WORD19`
#> [1] "100"
#> 
#> $`BD$WORD20`
#> [1] "100"
#> 
#> $`BD$WORD21`
#> [1] "100"
#> 
#> $`BD$WORD22`
#> [1] "100"
#> 
#> $`BD$WORD23`
#> [1] "1"
#> 
#> $`BD$WORD24`
#> [1] "1"
#> 
#> $`BD$WORD25`
#> [1] "0"
#> 
#> $`BD$WORD26`
#> [1] "0"
#> 
#> $`BD$WORD27`
#> [1] "0"
#> 
#> $`BD$WORD28`
#> [1] "136"
#> 
#> $`BD$WORD29`
#> [1] "77"
#> 
#> $`BD$WORD30`
#> [1] "52"
#> 
#> $`BD$WORD31`
#> [1] "52"
#> 
#> $`BD$WORD32`
#> [1] "52"
#> 
#> $`BD$WORD33`
#> [1] "52"
#> 
#> $`BD$WORD34`
#> [1] "6"
#> 
#> $`BD$WORD35`
#> [1] "300"
#> 
#> $`BD$WORD36`
#> [1] "0"
#> 
#> $`BD$WORD37`
#> [1] "191"
#> 
#> $`BD$WORD38`
#> [1] "280"
#> 
#> $`BD$WORD39`
#> [1] "3"
#> 
#> $`BD$WORD40`
#> [1] "2"
#> 
#> $`BD$WORD41`
#> [1] "100"
#> 
#> $`BD$WORD42`
#> [1] "100"
#> 
#> $`BD$WORD43`
#> [1] "0"
#> 
#> $`BD$WORD44`
#> [1] "1023"
#> 
#> $`BD$WORD45`
#> [1] "1023"
#> 
#> $`BD$WORD46`
#> [1] "1023"
#> 
#> $`BD$WORD47`
#> [1] "53"
#> 
#> $`BD$WORD48`
#> [1] "721"
#> 
#> $`BD$WORD49`
#> [1] "127"
#> 
#> $`BD$WORD50`
#> [1] "22"
#> 
#> $`BD$WORD51`
#> [1] "52"
#> 
#> $`BD$WORD52`
#> [1] "0"
#> 
#> $`BD$WORD53`
#> [1] "0"
#> 
#> $`BD$WORD54`
#> [1] "0"
#> 
#> $`BD$WORD55`
#> [1] "0"
#> 
#> $`BD$WORD56`
#> [1] "0"
#> 
#> $`BD$WORD57`
#> [1] "0"
#> 
#> $`BD$WORD58`
#> [1] "0"
#> 
#> $`BD$WORD59`
#> [1] "0"
#> 
#> $`BD$WORD60`
#> [1] "0"
#> 
#> $`BD$WORD61`
#> [1] "0"
#> 
#> $`BD$WORD62`
#> [1] "0"
#> 
#> $`BD$WORD63`
#> [1] "0"
#> 
#> $`BD$LASERMODE`
#> [1] "1"
#> 
#> $CalibFile
#> [1] "FALSE"
#> 
#> $P7THRESVOL
#> [1] "52"
#> 
#> $`$FIL`
#> [1] "0877408774.B08"
#> 
#> $`$DATE`
#> [1] "03-Feb-06"
#> 
#> $`NUMBER WELL INFO KEYWORDS`
#> [1] "0"
#> 
#> $`$P5S`
#> character(0)
#> 
#> $`$P7S`
#> character(0)
#> 
#> $`&1Sample Vol`
#> [1] "200"
#> 
#> $`&2Number of Washes`
#> [1] "1"
#> 
#> $`&3Mixing Vol`
#> [1] "100"
#> 
#> $`&4Number of Mixes`
#> [1] "1"
#> 
#> $`&5Data File Prefix Part #1`
#> [1] "08774"
#> 
#> $`&6Data File Prefix Part #2`
#> [1] "08774"
#> 
#> $`&7Data File Prefix Part #3`
#> character(0)
#> 
#> $`&8Acquisition Doc.`
#> [1] "UDCP Acquisition Template"
#> 
#> $`&9Instr. Sett. File`
#> [1] "020306 InstrSettings"
#> 
#> $`&10Analysis Doc.`
#> character(0)
#> 
#> $FILENAME
#> [1] "/tmp/RtmpINjmPE/temp_libpath618f421a5c04/flowCore/extdata/0877408774.B08"
#> 
#> $transformation
#> [1] "applied"
#> 
#> $`flowCore_$P1Rmax`
#> [1] "1024"
#> 
#> $`flowCore_$P1Rmin`
#> [1] "0"
#> 
#> $`flowCore_$P2Rmax`
#> [1] "1024"
#> 
#> $`flowCore_$P2Rmin`
#> [1] "0"
#> 
#> $`flowCore_$P3Rmax`
#> [1] "10001"
#> 
#> $`flowCore_$P3Rmin`
#> [1] "1"
#> 
#> $`flowCore_$P4Rmax`
#> [1] "10001"
#> 
#> $`flowCore_$P4Rmin`
#> [1] "1"
#> 
#> $`flowCore_$P5Rmax`
#> [1] "10001"
#> 
#> $`flowCore_$P5Rmin`
#> [1] "1"
#> 
#> $`flowCore_$P6Rmax`
#> [1] "1024"
#> 
#> $`flowCore_$P6Rmin`
#> [1] "0"
#> 
#> $`flowCore_$P7Rmax`
#> [1] "10001"
#> 
#> $`flowCore_$P7Rmin`
#> [1] "1"
#> 
#> $`flowCore_$P8Rmax`
#> [1] "1024"
#> 
#> $`flowCore_$P8Rmin`
#> [1] "0"
#> 
#> $GUID
#> [1] "0877408774.B08"
#> 

keyword(samp, compact = TRUE)

#> $TIMETICKS
#> [1] "50"
#> 
#> $`SAMPLE ID`
#> [1] "Default Patient ID"
#> 
#> $`CASE NUMBER`
#> [1] "Default Case Number"
#> 
#> $CYTNUM
#> [1] "E5451"
#> 
#> $CalibFile
#> [1] "FALSE"
#> 
#> $`NUMBER WELL INFO KEYWORDS`
#> [1] "0"
#> 
#> $`&1Sample Vol`
#> [1] "200"
#> 
#> $`&2Number of Washes`
#> [1] "1"
#> 
#> $`&3Mixing Vol`
#> [1] "100"
#> 
#> $`&4Number of Mixes`
#> [1] "1"
#> 
#> $`&5Data File Prefix Part #1`
#> [1] "08774"
#> 
#> $`&6Data File Prefix Part #2`
#> [1] "08774"
#> 
#> $`&7Data File Prefix Part #3`
#> character(0)
#> 
#> $`&8Acquisition Doc.`
#> [1] "UDCP Acquisition Template"
#> 
#> $`&9Instr. Sett. File`
#> [1] "020306 InstrSettings"
#> 
#> $`&10Analysis Doc.`
#> character(0)
#> 
#> $FILENAME
#> [1] "/tmp/RtmpINjmPE/temp_libpath618f421a5c04/flowCore/extdata/0877408774.B08"
#> 
#> $transformation
#> [1] "applied"
#> 
#> $GUID
#> [1] "0877408774.B08"
#> 

keyword(samp, "FCSversion")

#> $FCSversion
#> [1] "2"
#> 

keyword(samp, function(x,...) paste(keyword(x, "SAMPLE ID"), keyword(x,
"GUID"), sep="_"))

#> [1] "Default Patient ID_0877408774.B08"

keyword(samp) <- list(foo="bar")

data(GvHD)
keyword(GvHD, list("GUID", cellnumber=function(x) nrow(x)))

#>        cellnumber
#> s5a01        3420
#> s5a02        3405
#> s5a03        3435
#> s5a04        8550
#> s5a05       10410
#> s5a06        3750
#> s5a07       13979
#> s6a01        2205
#> s6a02       13350
#> s6a03       11610
#> s6a04       13830
#> s6a05       13934
#> s6a06       19470
#> s6a07        9405
#> s7a01       16055
#> s7a02       10305
#> s7a03       30195
#> s7a04       59849
#> s7a05       39068
#> s7a06       11939
#> s7a07       13214
#> s9a01       12666
#> s9a02        8265
#> s9a03        4065
#> s9a04        7080
#> s9a05       66105
#> s9a06       49421
#> s9a07       12768
#> s10a01      17289
#> s10a02       4530
#> s10a03       6765
#> s10a04       9540
#> s10a05      25515
#> s10a06      24656
#> s10a07      25601

keyword(GvHD) <- list(sample=sampleNames(GvHD))