Class "kmeansFilter"

A filter that performs one-dimensional k-means (Lloyd-Max) clustering on a single flow parameter.

kmeansFilter(..., filterId="defaultKmeansFilter")

Arguments

...
kmeansFilter are defined by a single flow parameter and an associated list of k population names. They can be given as a character vector via a named argument, or as a list with a single named argument. In both cases the name will be used as the flow parameter and the content of the list or of the argument will be used as population names, after coercing to character. For example

kmeansFilter(FSC=c("a", "b", "c"))

or

kmeansFilter(list(SSC=1:3))

If the parameter is not fully realized, but instead is the result of a transformation operation, two arguments need to be passed to the constructor: the first one being the transform object and the second being a vector of population names which can be coerced to a character. For example

kmeansFilter(tf, c("D", "E"))

filterId
An optional parameter that sets the filterId of the object. The filter can later be identified by this name.

Value

Returns a kmeansFilter object for use in filtering flowFrames or other flow cytometry objects.

Details

The one-dimensional k-means filter is a multiple population filter capable of operating on a single flow parameter. It takes a parameter argument associated with two or more populations and results in the generation of an object of class multipleFilterResult. Populations are considered to be ordered such that the population with the smallest mean intensity will be the first population in the list and the population with the highest mean intensity will be the last population listed.

Slots

populations

Object of class character. The names of the k populations (or clusters) that will be created by the kmeansFilter. These names will later be used for the respective subpopulations in split operations and for the summary of the filterResult.

parameters

Object of class parameters, defining a single parameter for which the data in the flowFrame is to be clustered. This may also be a transformation object.

filterId

Object of class character, an identifier or name to reference the kmeansFilter object later on.

Note

See the documentation in the flowViz package for plotting of kmeansFilters.

Extends

Class parameterFilter, directly.

Class concreteFilter, by class parameterFilter, distance 2.

Class filter, by class parameterFilter, distance3.

Objects from the Class

Like all other filter objects in flowCore, kmeansFilter objects should be instantiated through their constructor kmeansFilter(). See the Usage section for details.

Methods

%in%

signature(x = "flowFrame", table = "kmeansFilter"): The workhorse used to evaluate the filter on data. Usage: This is usually not called directly by the user, but internally by the filter methods.

show

signature(object = "kmeansFilter"): Print information about the filter. Usage: The method is called automatically whenever the object is printed on the screen.

See also

Examples

## Loading example data
dat <- read.FCS(system.file("extdata","0877408774.B08",
package="flowCore"))

## Create the filter
kf <- kmeansFilter("FSC-H"=c("Pop1","Pop2","Pop3"), filterId="myKmFilter")

## Filtering using kmeansFilters
fres <- filter(dat, kf)
fres

#> A filterResult produced by the filter named 'myKmFilter'
#>  resulting in multiple populations:
#> 	 Pop1
#> 	 Pop2
#> 	 Pop3

summary(fres)

#> Pop1: 1236 of 10000 events (12.36%)
#> Pop2: 6996 of 10000 events (69.96%)
#> Pop3: 1768 of 10000 events (17.68%)

names(fres)

#> [1] "Pop1" "Pop2" "Pop3"

## The result of quadGate filtering are multiple sub-populations
## and we can split our data set accordingly
split(dat, fres)

#> $Pop1
#> flowFrame object '0877408774.B08 (Pop1)'
#> with 1236 cells and 8 observables:
#>      name              desc range minRange maxRange
#> $P1 FSC-H             FSC-H  1024 0.000000     1023
#> $P2 SSC-H             SSC-H  1024 0.000000     1023
#> $P3 FL1-H              <NA>  1024 1.009044    10000
#> $P4 FL2-H              <NA>  1024 1.009044    10000
#> $P5 FL3-H              <NA>  1024 1.009044    10000
#> $P6 FL1-A              <NA>  1024 0.000000     1023
#> $P7 FL4-H              <NA>  1024 1.009044    10000
#> $P8  Time Time (51.20 sec.)  1024 0.000000     1023
#> 165 keywords are stored in the 'description' slot
#> 
#> $Pop2
#> flowFrame object '0877408774.B08 (Pop2)'
#> with 6996 cells and 8 observables:
#>      name              desc range minRange maxRange
#> $P1 FSC-H             FSC-H  1024 0.000000     1023
#> $P2 SSC-H             SSC-H  1024 0.000000     1023
#> $P3 FL1-H              <NA>  1024 1.009044    10000
#> $P4 FL2-H              <NA>  1024 1.009044    10000
#> $P5 FL3-H              <NA>  1024 1.009044    10000
#> $P6 FL1-A              <NA>  1024 0.000000     1023
#> $P7 FL4-H              <NA>  1024 1.009044    10000
#> $P8  Time Time (51.20 sec.)  1024 0.000000     1023
#> 165 keywords are stored in the 'description' slot
#> 
#> $Pop3
#> flowFrame object '0877408774.B08 (Pop3)'
#> with 1768 cells and 8 observables:
#>      name              desc range minRange maxRange
#> $P1 FSC-H             FSC-H  1024 0.000000     1023
#> $P2 SSC-H             SSC-H  1024 0.000000     1023
#> $P3 FL1-H              <NA>  1024 1.009044    10000
#> $P4 FL2-H              <NA>  1024 1.009044    10000
#> $P5 FL3-H              <NA>  1024 1.009044    10000
#> $P6 FL1-A              <NA>  1024 0.000000     1023
#> $P7 FL4-H              <NA>  1024 1.009044    10000
#> $P8  Time Time (51.20 sec.)  1024 0.000000     1023
#> 165 keywords are stored in the 'description' slot
#> 

## We can limit the splitting to one or several sub-populations
split(dat, fres, population="Pop1")

#> $Pop1
#> flowFrame object '0877408774.B08 (Pop1)'
#> with 1236 cells and 8 observables:
#>      name              desc range minRange maxRange
#> $P1 FSC-H             FSC-H  1024 0.000000     1023
#> $P2 SSC-H             SSC-H  1024 0.000000     1023
#> $P3 FL1-H              <NA>  1024 1.009044    10000
#> $P4 FL2-H              <NA>  1024 1.009044    10000
#> $P5 FL3-H              <NA>  1024 1.009044    10000
#> $P6 FL1-A              <NA>  1024 0.000000     1023
#> $P7 FL4-H              <NA>  1024 1.009044    10000
#> $P8  Time Time (51.20 sec.)  1024 0.000000     1023
#> 165 keywords are stored in the 'description' slot
#> 

split(dat, fres, population=list(keep=c("Pop1","Pop2")))

#> $keep
#> flowFrame object '0877408774.B08 (Pop1,Pop2)'
#> with 8232 cells and 8 observables:
#>      name              desc range minRange maxRange
#> $P1 FSC-H             FSC-H  1024 0.000000     1023
#> $P2 SSC-H             SSC-H  1024 0.000000     1023
#> $P3 FL1-H              <NA>  1024 1.009044    10000
#> $P4 FL2-H              <NA>  1024 1.009044    10000
#> $P5 FL3-H              <NA>  1024 1.009044    10000
#> $P6 FL1-A              <NA>  1024 0.000000     1023
#> $P7 FL4-H              <NA>  1024 1.009044    10000
#> $P8  Time Time (51.20 sec.)  1024 0.000000     1023
#> 3 keywords are stored in the 'description' slot
#>