Class "manyFilterResult"

The result of a several related, but possibly overlapping filter results. The usual creator of this object will usually be a filter operation on a flowFrame object.

Slots

subSet

Object of class "matrix".

frameId

Object of class "character" referencing the flowFrame object filtered. Used for sanity checking.

filterDetails

Object of class "list" describing the filter applied.

filterId

Object of class "character" referencing the filter applied.

dependency

Any dependencies between the filters. Currently not used.

Extends

Class "filterResult", directly. Class "filter", by class "filterResult", distance 2.

Methods

[, [[

subsetting. If x is manyFilterResult, then x[[i]] a filterResult object. The semantics is similar to the behavior of the subsetting operators for lists.

length

number of filterResult objects in the set.

names

names of the filterResult objects in the set.

summary

summary filterResult objects in the set.

See also

Examples

showClass("manyFilterResult")

#> Class "manyFilterResult" [package "flowCore"]
#> 
#> Slots:
#>                                                                             
#> Name:         subSet    dependency       frameId filterDetails      filterId
#> Class:        matrix           ANY     character          list     character
#> 
#> Extends: 
#> Class "filterResult", directly
#> Class "concreteFilter", by class "filterResult", distance 2
#> Class "filter", by class "filterResult", distance 3