get or update the marker names

marker names corresponds to the 'desc' column of the phenoData of the flowFrame.

markernames(object, ...)

# S4 method for flowFrame
markernames(object)

markernames(object) <- value

# S4 method for flowFrame
markernames(object) <- value

# S4 method for flowSet
markernames(object)

# S4 method for flowSet
markernames(object) <- value

Arguments

object
flowFrame or flowSet

...
not used

value
a named list or character vector. the names corresponds to the name(channel) and actual values are the desc(marker).

Value

marker names as a character vector. The marker names for FSC,SSC and Time channels are automatically excluded in the returned value. When object is a flowSet and the marker names are not consistent across flowFrames, it returns a list of unique marker sets.

Details

When extract marker names from a flowSet, it throws the warning if the marker names are not all the same across samples.

Examples

data(GvHD)
fr <- GvHD[[1]]
markernames(fr)

#>        FL1-H        FL2-H        FL3-H        FL4-H 
#>  "CD15 FITC"    "CD45 PE" "CD14 PerCP"   "CD33 APC" 

chnls <- c("FL1-H", "FL3-H")
markers <- c("CD15", "CD14")
names(markers) <- chnls
markernames(fr) <- markers
markernames(fr)

#>      FL1-H      FL2-H      FL3-H      FL4-H 
#>     "CD15"  "CD45 PE"     "CD14" "CD33 APC" 

fs <- GvHD[1:3]
markernames(fs)

#>        FL1-H        FL2-H        FL3-H        FL4-H 
#>  "CD15 FITC"    "CD45 PE" "CD14 PerCP"   "CD33 APC"