Obtain information about parameters for flow cytometry objects.

Many different objects in flowCore are associated with one or more parameters. This includes filter, flowFrame and parameterFilter objects that all either describe or use parameters.

parameters(object, ...)

Arguments

object
Object of class filter, flowFrame or parameterFilter.

...
Further arguments that get passed on to the methods.

Value

When applied to a flowFrame object, the result is an AnnotatedDataFrame describing the parameters recorded by the cytometer. For other objects it will usually return a vector of names used by the object for its calculations.

Methods

parameters(object = "filter")

Returns for all objects that inherit from filter a vector of parameters on which a gate is defined.

parameters(object = "parameterFilter")

see above

parameters(object = "setOperationFilter")

see above

parameters(object = "filterReference")

see above

parameters(object = "flowFrame")

Returns an AnnotatedDataFrame containing detailed descriptions about the measurement parameters of the flowFrame. For flowFrame objects there also exists a replacement method.

Examples

 samp <- read.FCS(system.file("extdata","0877408774.B08", package="flowCore"))
 parameters(samp)

#> An object of class 'AnnotatedDataFrame'
#>   rowNames: $P1 $P2 ... $P8 (8 total)
#>   varLabels: name desc ... maxRange (5 total)
#>   varMetadata: labelDescription

 print(samp@parameters@data)

#>      name              desc range minRange maxRange
#> $P1 FSC-H             FSC-H  1024 0.000000     1023
#> $P2 SSC-H             SSC-H  1024 0.000000     1023
#> $P3 FL1-H              <NA>  1024 1.009044    10000
#> $P4 FL2-H              <NA>  1024 1.009044    10000
#> $P5 FL3-H              <NA>  1024 1.009044    10000
#> $P6 FL1-A              <NA>  1024 0.000000     1023
#> $P7 FL4-H              <NA>  1024 1.009044    10000
#> $P8  Time Time (51.20 sec.)  1024 0.000000     1023