Read the TEXT section of a FCS file

Read (part of) the TEXT section of a Data File Standard for Flow Cytometry that contains FACS keywords.

read.FCSheader(files, path = ".", keyword = NULL, ...)

Arguments

files
Character vector of filenames.

path
Directory where to look for the files.

keyword
An optional character vector that specifies the FCS keyword to read.

...
other arguments passed to link[flowCore]{read.FCS}

Value

A list of character vectors. Each element of the list correspond to one FCS file.

Details

The function read.FCSheader works with the output of the FACS machine software from a number of vendors (FCS 2.0, FCS 3.0 and List Mode Data LMD). The output of the function is the TEXT section of the FCS files. The user can specify some keywords to limit the output to the information of interest.

See also

Examples

samp <- read.FCSheader(system.file("extdata",
   "0877408774.B08", package="flowCore"))
samp

#> $`/tmp/RtmpINjmPE/temp_libpath618f421a5c04/flowCore/extdata/0877408774.B08`
#>                        FCSversion                          $BYTEORD 
#>                               "2"                         "4,3,2,1" 
#>                         $DATATYPE                         $NEXTDATA 
#>                               "I"                               "0" 
#>                              $SYS                           CREATOR 
#> "Macintosh System Software 9.2.2"               "CELLQuest<aa> 3.3" 
#>                              $TOT                             $MODE 
#>                           "10000"                               "L" 
#>                              $PAR                              $P1N 
#>                               "8"                           "FSC-H" 
#>                              $P1R                              $P1B 
#>                            "1024"                              "16" 
#>                              $P1E                              $P2N 
#>                             "0,0"                           "SSC-H" 
#>                              $P2R                              $P2B 
#>                            "1024"                              "16" 
#>                              $P2E                              $P3N 
#>                             "0,0"                           "FL1-H" 
#>                              $P3R                              $P3B 
#>                            "1024"                              "16" 
#>                              $P3E                              $P4N 
#>                             "4,1"                           "FL2-H" 
#>                              $P4R                              $P4B 
#>                            "1024"                              "16" 
#>                              $P4E                              $P5N 
#>                             "4,1"                           "FL3-H" 
#>                              $P5R                              $P5B 
#>                            "1024"                              "16" 
#>                              $P5E                              $P1S 
#>                             "4,1"                           "FSC-H" 
#>                              $P2S                              $P3S 
#>                           "SSC-H"                                "" 
#>                              $P4S                              $P6N 
#>                                ""                           "FL1-A" 
#>                              $P6R                              $P6B 
#>                            "1024"                              "16" 
#>                              $P6E                         TIMETICKS 
#>                             "0,0"                              "50" 
#>                              $P7N                              $P7R 
#>                           "FL4-H"                            "1024" 
#>                              $P7E                              $P7B 
#>                             "4,1"                              "16" 
#>                              $P8N                              $P8R 
#>                            "Time"                            "1024" 
#>                              $P8E                              $P8B 
#>                             "0,0"                              "16" 
#>                              $P8S                         SAMPLE ID 
#>               "Time (51.20 sec.)"              "Default Patient ID" 
#>                              $SRC                       CASE NUMBER 
#>            "Default Patient Name"             "Default Case Number" 
#>                              $CYT                            CYTNUM 
#>                     "FACSCalibur"                           "E5451" 
#>                             $BTIM                             $ETIM 
#>                        "13:13:54"                        "13:14:27" 
#>                  BD$AcqLibVersion                           BD$NPAR 
#>                             "3.1"                               "7" 
#>                            BD$P1N                            BD$P2N 
#>                           "FSC-H"                           "SSC-H" 
#>                            BD$P3N                            BD$P4N 
#>                           "FL1-H"                           "FL2-H" 
#>                            BD$P5N                            BD$P6N 
#>                           "FL3-H"                           "FL1-A" 
#>                            BD$P7N                          BD$WORD0 
#>                           "FL4-H"                              "24" 
#>                          BD$WORD1                          BD$WORD2 
#>                             "474"                             "594" 
#>                          BD$WORD3                          BD$WORD4 
#>                             "530"                             "700" 
#>                          BD$WORD5                          BD$WORD6 
#>                             "402"                             "401" 
#>                          BD$WORD7                          BD$WORD8 
#>                             "401"                             "402" 
#>                          BD$WORD9                         BD$WORD10 
#>                             "401"                             "299" 
#>                         BD$WORD11                         BD$WORD12 
#>                             "287"                             "557" 
#>                         BD$WORD13                         BD$WORD14 
#>                               "0"                             "483" 
#>                         BD$WORD15                         BD$WORD16 
#>                             "612"                             "546" 
#>                         BD$WORD17                         BD$WORD18 
#>                             "724"                             "104" 
#>                         BD$WORD19                         BD$WORD20 
#>                             "100"                             "100" 
#>                         BD$WORD21                         BD$WORD22 
#>                             "100"                             "100" 
#>                         BD$WORD23                         BD$WORD24 
#>                               "1"                               "1" 
#>                         BD$WORD25                         BD$WORD26 
#>                               "0"                               "0" 
#>                         BD$WORD27                         BD$WORD28 
#>                               "0"                             "136" 
#>                         BD$WORD29                         BD$WORD30 
#>                              "77"                              "52" 
#>                         BD$WORD31                         BD$WORD32 
#>                              "52"                              "52" 
#>                         BD$WORD33                         BD$WORD34 
#>                              "52"                               "6" 
#>                         BD$WORD35                         BD$WORD36 
#>                             "300"                               "0" 
#>                         BD$WORD37                         BD$WORD38 
#>                             "191"                             "280" 
#>                         BD$WORD39                         BD$WORD40 
#>                               "3"                               "2" 
#>                         BD$WORD41                         BD$WORD42 
#>                             "100"                             "100" 
#>                         BD$WORD43                         BD$WORD44 
#>                               "0"                            "1023" 
#>                         BD$WORD45                         BD$WORD46 
#>                            "1023"                            "1023" 
#>                         BD$WORD47                         BD$WORD48 
#>                              "53"                             "721" 
#>                         BD$WORD49                         BD$WORD50 
#>                             "127"                              "22" 
#>                         BD$WORD51                         BD$WORD52 
#>                              "52"                               "0" 
#>                         BD$WORD53                         BD$WORD54 
#>                               "0"                               "0" 
#>                         BD$WORD55                         BD$WORD56 
#>                               "0"                               "0" 
#>                         BD$WORD57                         BD$WORD58 
#>                               "0"                               "0" 
#>                         BD$WORD59                         BD$WORD60 
#>                               "0"                               "0" 
#>                         BD$WORD61                         BD$WORD62 
#>                               "0"                               "0" 
#>                         BD$WORD63                      BD$LASERMODE 
#>                               "0"                               "1" 
#>                         CalibFile                        P7THRESVOL 
#>                           "FALSE"                              "52" 
#>                              $FIL                             $DATE 
#>                  "0877408774.B08"                       "03-Feb-06" 
#>         NUMBER WELL INFO KEYWORDS                              $P5S 
#>                               "0"                                "" 
#>                              $P7S                      &1Sample Vol 
#>                                ""                             "200" 
#>                &2Number of Washes                      &3Mixing Vol 
#>                               "1"                             "100" 
#>                 &4Number of Mixes        &5Data File Prefix Part #1 
#>                               "1"                           "08774" 
#>        &6Data File Prefix Part #2        &7Data File Prefix Part #3 
#>                           "08774"                                "" 
#>                &8Acquisition Doc.               &9Instr. Sett. File 
#>       "UDCP Acquisition Template"            "020306 InstrSettings" 
#>                  &10Analysis Doc. 
#>                                "" 
#> 

samp <- read.FCSheader(system.file("extdata",
   "0877408774.B08", package="flowCore"), keyword=c("$DATE", "$FIL"))
samp

#> $`/tmp/RtmpINjmPE/temp_libpath618f421a5c04/flowCore/extdata/0877408774.B08`
#>            $DATE             $FIL 
#>      "03-Feb-06" "0877408774.B08" 
#>