Create the definition of a scale transformation function to be applied on a data set

Create the definition of the scale Transformation that will be applied on some parameter via the transform method. The definition of this function is currently x = (x-a)/(b-a). The transformation would normally be used to convert to a 0-1 scale. In this case, b would be the maximum possible value and a would be the minimum possible value.

scaleTransform(transformationId="defaultScaleTransform", a, b)

Arguments

transformationId
character string to identify the transformation

a
double that corresponds to the value that will be transformed to 0

b
double that corresponds to the value that will be transformed to 1

Value

Returns an object of class transform.

See also

Examples

samp <- read.FCS(system.file("extdata",
   "0877408774.B08", package="flowCore"))
  scaleTrans <- scaleTransform(transformationId="Truncate-transformation", a=1, b=10^4)
  dataTransform <- transform(samp, transformList('FSC-H', scaleTrans))