Class "transformList"

A list of transformMaps to be applied to a list of parameters.

transformList(from, tfun, to=from, transformationId =
"defaultTransformation")

Arguments

from, to
Characters giving the names of the measurement parameter on which to transform on and into which the result is supposed to be stored. If both are equal, the existing parameters will be overwritten.

tfun
A list if functions or a character vector of the names of the functions used to transform the data. R's recycling rules apply, so a single function can be given to be used on all parameters.

transformationId
The identifier for the object.

Slots

transforms

Object of class "list", where each list item is of class transformMap.

transformationId

Object of class "character", the identifier for the object.

Objects from the Class

Objects can be created by calls of the form new("transformList", ...), by calling the transform method with key-value pair arguments of the form key equals character and value equals function, or by using the constructor transformList. See below for details

Methods

colnames

signature(x = "transformList"): This returns the names of the parameters that are to be transformed.

c

signature(x = "transformList"): Concatenate transformLists or regular lists and transformLists.

%on%

signature(e1 = "transformList", e2 = "flowFrame"): Perform a transformation using the transformList on a flowFrame or flowSet.

See also

Examples

tl <- transformList(c("FSC-H", "SSC-H"), list(log, asinh))
colnames(tl)

#> [1] "FSC-H" "SSC-H"

c(tl, transformList("FL1-H", "linearTransform"))

#> An object of class "transformList"
#> Slot "transforms":
#> $`FSC-H`
#> transformMap for parameter 'FSC-H' mapping to 'FSC-H'
#> 
#> $`SSC-H`
#> transformMap for parameter 'SSC-H' mapping to 'SSC-H'
#> 
#> $`FL1-H`
#> transformMap for parameter 'FL1-H' mapping to 'FL1-H'
#> 
#> 
#> Slot "transformationId":
#> [1] "defaultTransformation"
#> 

data(GvHD)
transform(GvHD[[1]], tl)

#> flowFrame object 's5a01'
#> with 3420 cells and 8 observables:
#>      name              desc range minRange     maxRange
#> $P1 FSC-H        FSC-Height  1024     -Inf     6.930495
#> $P2 SSC-H        SSC-Height  1024        0     7.623642
#> $P3 FL1-H         CD15 FITC  1024        1 10000.000000
#> $P4 FL2-H           CD45 PE  1024        1 10000.000000
#> $P5 FL3-H        CD14 PerCP  1024        1 10000.000000
#> $P6 FL2-A              <NA>  1024        0  1023.000000
#> $P7 FL4-H          CD33 APC  1024        1 10000.000000
#> $P8  Time Time (51.20 sec.)  1024        0  1023.000000
#> 169 keywords are stored in the 'description' slot