A class for mapping transforms between parameters

This class provides a mapping between parameters and transformed parameters via a function.

Slots

output

Name of the transformed parameter.

input

Name of the parameter to transform.

f

Function used to accomplish the transform.

Objects from the Class

Objects of this type are not usually created by the user, except perhaps in special circumstances. They are generally automatically created by the inline transform process during the creation of a transformFilter, or by a call to the transformList constructor.

Methods

show

signature(object = "transformList"): Print details about the object.

See also

Examples

new("transformMap", input="FSC-H", output="FSC-H", f=log)

#> transformMap for parameter 'FSC-H' mapping to 'FSC-H'