Class "unitytransform"

Unity transform class transforms parameters names provided as characters into unity transform objects which can be evaluated to retrieve the corresponding columns from the data frame

Slots

.Data

Object of class "function".

parameters

Object of class "character" -- the flow parameters to be transformed.

transformationId

Object of class "character" -- a unique Id to reference the transformation.

Objects from the Class

Objects can be created by calls to the constructor unitytransform(parameters,transformationId).

Extends

Class "transform", directly.

Class "transformation", by class "transform", distance 2.

Class "characterOrTransformation", by class "transform", distance 3.

See also

Examples

  dat <- read.FCS(system.file("extdata","0877408774.B08",
  package="flowCore"))
  un1<-unitytransform(c("FSC-H","SSC-H"),transformationId="un1")
  transOut<-eval(un1)(exprs(dat))