Write an FCS file

Write FCS file from a flowFrame

write.FCS(x, filename, what="numeric", delimiter = "|", endian="big")

Arguments

x
A flowFrame.

filename
A character scalar giving the output file name.

what
A character scalar defining the output data type. One in integer, numeric, double. Note that forcing the data type to integer may result in considerable loss of precision if the data has been transformed. We recommend using the default data type unless disc space is an issue.

delimiter
a single character to separate the FCS keyword/value pairs. Default is : "|"

endian
a character, either "little" or "big" (default), specifying the most significant or least significant byte is stored first in a 32 bit word.

Value

A character vector of the file name.

Details

The function write.FCS creates FCS 3.0 standard file from an object of class flowFrame.

For specifications of FCS 3.0 see http://www.isac-net.org and the file ../doc/fcs3.html in the doc directory of the package.

See also

Examples

## a sample file
inFile <- system.file("extdata", "0877408774.B08", package="flowCore")
foo <- read.FCS(inFile, transform=FALSE)
outFile <- file.path(tempdir(), "foo.fcs")

## now write out into a file
write.FCS(foo, outFile)

#> [1] "/tmp/RtmprOujRC/foo.fcs"

bar <- read.FCS(outFile, transform=FALSE)
all(exprs(foo) == exprs(bar))

#> [1] TRUE