Write an FCS file

Write FCS file for each flowFrame in a flowSet

write.flowSet(x, outdir=identifier(x), filename, ...)

Arguments

x
A flowSet.

outdir
A character scalar giving the output directory. As the default, the output of identifier(x) is used.

filename
A character scalar or vector giving the output file names. By default, the function will use the identifiers of the individual flowFrames as the file name, potentially adding the .fcs suffix unless a file extension is already present. Alternatively, one can supply either a character scalar, in which case the prefix i_ is appended (i being an integer in seq_len(length(x))), or a character vector of the same length as the flowSet x.

...
Further arguments that are passed on to write.FCS.

Value

A character vector of the output directory.

Details

The function write.flowSet creates FCS 3.0 standard file for all flowFrames in an object of class flowSet. In addition, it will write the content of the phenoData slot in the ASCII file "annotation.txt". This file can subsequently be used to reconstruct the whole flowSet using the read.flowSet function, e.g.:

read.flowSet(path=outdir, phenoData="annotation.txt"

The function uses write.FCS for the actual writing of the FCS files.

See also

Examples

## sample data
data(GvHD)
foo <- GvHD[1:5]
outDir <- file.path(tempdir(), "foo")

## now write out into  files
write.flowSet(foo, outDir)

#> [1] "/tmp/RtmpexaeiJ/foo"

dir(outDir)

#> [1] "annotation.txt" "s5a01.fcs"      "s5a02.fcs"      "s5a03.fcs"     
#> [5] "s5a04.fcs"      "s5a05.fcs"     

## and read back in
bar <- read.flowSet(path=outDir, phenoData="annotation.txt")