Coerce the list of the keywords into a character Also flatten spillover matrix into a string

Coerce the list of the keywords into a character Also flatten spillover matrix into a string

collapse_desc(d, collapse.spill = TRUE)

Arguments

d
a named list of keywords

collapse.spill
whether to flatten spillover matrix to a string

Value

a list of strings

Examples

data(GvHD)
fr <- GvHD[[1]]
collapse_desc(keyword(fr))

#> $`&10pt name`
#> [1] "BJ#230500"
#> 
#> $`&11Day`
#> [1] "-6d"
#> 
#> $`&12Sample ID`
#> [1] "Leukocytes"
#> 
#> $`&13Analysis Doc.`
#> [1] " "
#> 
#> $`&1Sample Vol`
#> [1] "150"
#> 
#> $`&2Number of Washes`
#> [1] "1"
#> 
#> $`&3Mixing Vol`
#> [1] "50"
#> 
#> $`&4Number of Mixes`
#> [1] "1"
#> 
#> $`&5Data File Prefix Part #1`
#> [1] " "
#> 
#> $`&6Data File Prefix Part #2`
#> [1] " "
#> 
#> $`&7Data File Prefix Part #3`
#> [1] " "
#> 
#> $`&8Acquisition Doc.`
#> [1] "Lymph"    "4"        "color"    "aq"       "template"
#> 
#> $`&9Instr. Sett. File`
#> [1] "E#5"      "Settings"
#> 
#> $`$BTIM`
#> [1] "09:51:34"
#> 
#> $`$BYTEORD`
#> [1] "4,3,2,1"
#> 
#> $`$CYT`
#> [1] "FACSCalibur"
#> 
#> $`$DATATYPE`
#> [1] "I"
#> 
#> $`$DATE`
#> [1] "20-Aug-02"
#> 
#> $`$ETIM`
#> [1] "09:52:16"
#> 
#> $`$FIL`
#> [1] "A01"
#> 
#> $`$MODE`
#> [1] "L"
#> 
#> $`$NEXTDATA`
#> [1] "0"
#> 
#> $`$P1B`
#> [1] "16"
#> 
#> $`$P1E`
#> [1] "0,0"
#> 
#> $`$P1N`
#> [1] "FSC-H"
#> 
#> $`$P1R`
#> [1] "1024"
#> 
#> $`$P1S`
#> [1] "FSC-Height"
#> 
#> $`$P2B`
#> [1] "16"
#> 
#> $`$P2E`
#> [1] "0,0"
#> 
#> $`$P2N`
#> [1] "SSC-H"
#> 
#> $`$P2R`
#> [1] "1024"
#> 
#> $`$P2S`
#> [1] "SSC-Height"
#> 
#> $`$P3B`
#> [1] "16"
#> 
#> $`$P3E`
#> [1] "4,0"
#> 
#> $`$P3N`
#> [1] "FL1-H"
#> 
#> $`$P3R`
#> [1] "1024"
#> 
#> $`$P3S`
#> [1] "CD15" "FITC"
#> 
#> $`$P4B`
#> [1] "16"
#> 
#> $`$P4E`
#> [1] "4,0"
#> 
#> $`$P4N`
#> [1] "FL2-H"
#> 
#> $`$P4R`
#> [1] "1024"
#> 
#> $`$P4S`
#> [1] "CD45" "PE"  
#> 
#> $`$P5B`
#> [1] "16"
#> 
#> $`$P5E`
#> [1] "4,0"
#> 
#> $`$P5N`
#> [1] "FL3-H"
#> 
#> $`$P5R`
#> [1] "1024"
#> 
#> $`$P5S`
#> [1] "CD14"  "PerCP"
#> 
#> $`$P6B`
#> [1] "16"
#> 
#> $`$P6E`
#> [1] "0,0"
#> 
#> $`$P6N`
#> [1] "FL2-A"
#> 
#> $`$P6R`
#> [1] "1024"
#> 
#> $`$P7B`
#> [1] "16"
#> 
#> $`$P7E`
#> [1] "4,0"
#> 
#> $`$P7N`
#> [1] "FL4-H"
#> 
#> $`$P7R`
#> [1] "1024"
#> 
#> $`$P7S`
#> [1] "CD33" "APC" 
#> 
#> $`$P8B`
#> [1] "16"
#> 
#> $`$P8E`
#> [1] "0,0"
#> 
#> $`$P8N`
#> [1] "Time"
#> 
#> $`$P8R`
#> [1] "1024"
#> 
#> $`$P8S`
#> [1] "Time"   "(51.20" "sec.)" 
#> 
#> $`$PAR`
#> [1] "8"
#> 
#> $`$SRC`
#> [1] "Default" "Patient" "Name"   
#> 
#> $`$SYS`
#> [1] "Macintosh" "System"    "Software"  "9.0.4"    
#> 
#> $`$TOT`
#> [1] "3420"
#> 
#> $`BD$AcqLibVersion`
#> [1] "3.1"
#> 
#> $`BD$LASERMODE`
#> [1] "1"
#> 
#> $`BD$NPAR`
#> [1] "7"
#> 
#> $`BD$P1N`
#> [1] "FSC-H"
#> 
#> $`BD$P2N`
#> [1] "SSC-H"
#> 
#> $`BD$P3N`
#> [1] "FL1-H"
#> 
#> $`BD$P4N`
#> [1] "FL2-H"
#> 
#> $`BD$P5N`
#> [1] "FL3-H"
#> 
#> $`BD$P6N`
#> [1] "FL2-A"
#> 
#> $`BD$P7N`
#> [1] "FL4-H"
#> 
#> $`BD$WORD0`
#> [1] "24"
#> 
#> $`BD$WORD1`
#> [1] "440"
#> 
#> $`BD$WORD10`
#> [1] "300"
#> 
#> $`BD$WORD11`
#> [1] "291"
#> 
#> $`BD$WORD12`
#> [1] "583"
#> 
#> $`BD$WORD13`
#> [1] "0"
#> 
#> $`BD$WORD14`
#> [1] "449"
#> 
#> $`BD$WORD15`
#> [1] "583"
#> 
#> $`BD$WORD16`
#> [1] "519"
#> 
#> $`BD$WORD17`
#> [1] "585"
#> 
#> $`BD$WORD18`
#> [1] "133"
#> 
#> $`BD$WORD19`
#> [1] "100"
#> 
#> $`BD$WORD2`
#> [1] "562"
#> 
#> $`BD$WORD20`
#> [1] "100"
#> 
#> $`BD$WORD21`
#> [1] "100"
#> 
#> $`BD$WORD22`
#> [1] "100"
#> 
#> $`BD$WORD23`
#> [1] "1"
#> 
#> $`BD$WORD24`
#> [1] "1"
#> 
#> $`BD$WORD25`
#> [1] "0"
#> 
#> $`BD$WORD26`
#> [1] "0"
#> 
#> $`BD$WORD27`
#> [1] "0"
#> 
#> $`BD$WORD28`
#> [1] "136"
#> 
#> $`BD$WORD29`
#> [1] "52"
#> 
#> $`BD$WORD3`
#> [1] "515"
#> 
#> $`BD$WORD30`
#> [1] "52"
#> 
#> $`BD$WORD31`
#> [1] "52"
#> 
#> $`BD$WORD32`
#> [1] "52"
#> 
#> $`BD$WORD33`
#> [1] "52"
#> 
#> $`BD$WORD34`
#> [1] "4"
#> 
#> $`BD$WORD35`
#> [1] "257"
#> 
#> $`BD$WORD36`
#> [1] "1"
#> 
#> $`BD$WORD37`
#> [1] "119"
#> 
#> $`BD$WORD38`
#> [1] "280"
#> 
#> $`BD$WORD39`
#> [1] "3"
#> 
#> $`BD$WORD4`
#> [1] "567"
#> 
#> $`BD$WORD40`
#> [1] "3"
#> 
#> $`BD$WORD41`
#> [1] "100"
#> 
#> $`BD$WORD42`
#> [1] "100"
#> 
#> $`BD$WORD43`
#> [1] "0"
#> 
#> $`BD$WORD44`
#> [1] "1023"
#> 
#> $`BD$WORD45`
#> [1] "1023"
#> 
#> $`BD$WORD46`
#> [1] "1023"
#> 
#> $`BD$WORD47`
#> [1] "53"
#> 
#> $`BD$WORD48`
#> [1] "637"
#> 
#> $`BD$WORD49`
#> [1] "42"
#> 
#> $`BD$WORD5`
#> [1] "405"
#> 
#> $`BD$WORD50`
#> [1] "29"
#> 
#> $`BD$WORD51`
#> [1] "52"
#> 
#> $`BD$WORD52`
#> [1] "0"
#> 
#> $`BD$WORD53`
#> [1] "0"
#> 
#> $`BD$WORD54`
#> [1] "0"
#> 
#> $`BD$WORD55`
#> [1] "0"
#> 
#> $`BD$WORD56`
#> [1] "0"
#> 
#> $`BD$WORD57`
#> [1] "0"
#> 
#> $`BD$WORD58`
#> [1] "0"
#> 
#> $`BD$WORD59`
#> [1] "0"
#> 
#> $`BD$WORD6`
#> [1] "405"
#> 
#> $`BD$WORD60`
#> [1] "0"
#> 
#> $`BD$WORD61`
#> [1] "0"
#> 
#> $`BD$WORD62`
#> [1] "0"
#> 
#> $`BD$WORD63`
#> [1] "0"
#> 
#> $`BD$WORD7`
#> [1] "406"
#> 
#> $`BD$WORD8`
#> [1] "405"
#> 
#> $`BD$WORD9`
#> [1] "406"
#> 
#> $CalibFile
#> [1] "FALSE"
#> 
#> $`CASE NUMBER`
#> [1] "Default" "Case"    "Number" 
#> 
#> $CREATOR
#> [1] "CELLQuest\xaa" "3.3"          
#> 
#> $CYTNUM
#> [1] "E3597"
#> 
#> $`END BATCH`
#> [1] "FALSE"
#> 
#> $FCSversion
#> [1] "2"
#> 
#> $FILENAME
#> [1] "s5a01"
#> 
#> $`NUMBER WELL INFO KEYWORDS`
#> [1] "3"
#> 
#> $P7THRESVOL
#> [1] "52"
#> 
#> $`PANEL NAME`
#> [1] "nooooo"
#> 
#> $`SAMPLE ID`
#> [1] "Default" "Patient" "ID"     
#> 
#> $`START BATCH`
#> [1] "FALSE"
#> 
#> $TIMETICKS
#> [1] "50"
#> 
#> $transformation
#> [1] "applied"
#>