Class "filterResultList"

Container to store the result of applying a filter on a flowSet object

Slots

.Data

Object of class "list". The class directly extends list, and this slot holds the list data.

frameId

Object of class "character" The IDs of the flowFrames in the filtered flowSet.

filterDetails

Object of class "list". Since filterResultList inherits from filterResult, this slot has to be set. It contains only the input filter.

filterId

Object of class "character". The identifier for the object.

Objects from the Class

Objects are created by applying a filter on a flowSet. The user doesn't have to deal with manual object instantiation.

Extends

Class "list", from data part. Class "filterResult", directly. Class "concreteFilter", by class "filterResult", distance 2. Class "filter", by class "filterResult", distance 3.

Methods

[

signature(x = "filterResultList", i = "ANY"): Subset to filterResultList.

[[

signature(x = "filterResultList", i = "ANY"): Subset to individual filterResult.

names

signature(x = "filterResultList"): Accessor to the frameId slot.

parameters

signature(object = "filterResultList"): Return parameters on which data has been filtered.

show

signature(object = "filterResultList"): Print details about the object.

split

signature(x = "flowSet", f = "filterResultList"): Split a flowSet based on the results in the filterResultlIst. See split for details.

summary

signature(object = "filterResultList"): Summarize the filtering operation. This creates a filterSummaryList object.

See also

Examples

library(flowStats)
## Loading example data and creating a curv1Filter
data(GvHD)
dat <- GvHD[1:3]
c1f <- curv1Filter(filterId="myCurv1Filter", x=list("FSC-H"), bwFac=2)

## applying the filter
fres <- filter(dat, c1f)
fres

#> A list of filterResults for a flowSet containing 3 frames
#> produced by the filter named 'myCurv1Filter'

## subsetting the list
fres[[1]]

#> A filterResult produced by the filter named 'myCurv1Filter'
#>  resulting in multiple populations:
#> 	 rest
#> 	 peak 1

fres[1:2]

#> A list of filterResults for a flowSet containing 2 frames
#> produced by the filter named 'myCurv1Filter'

## details about the object
parameters(fres)

#> $s5a01
#> [1] "FSC-H"
#> 
#> $s5a02
#> [1] "FSC-H"
#> 
#> $s5a03
#> [1] "FSC-H"
#> 

names(fres)

#> [1] "s5a01" "s5a02" "s5a03"

summary(fres)

#> filter summary for frame 's5a01'
#>  rest: 1002 of 3420 events (29.30%)
#>  peak 1: 2418 of 3420 events (70.70%)
#> 
#> filter summary for frame 's5a02'
#>  rest: 1026 of 3405 events (30.13%)
#>  peak 1: 2379 of 3405 events (69.87%)
#> 
#> filter summary for frame 's5a03'
#>  rest: 369 of 3435 events (10.74%)
#>  peak 1: 3066 of 3435 events (89.26%)
#> 

## splitting based on the filterResults
split(dat, fres)

#> $rest
#> A flowSet with 3 experiments.
#> 
#> An object of class 'AnnotatedDataFrame'
#>   rowNames: s5a01 s5a02 s5a03
#>   varLabels: Patient Visit ... population (6 total)
#>   varMetadata: labelDescription
#> 
#>   column names:
#>   FSC-H SSC-H FL1-H FL2-H FL3-H FL2-A FL4-H Time
#> 
#> $`peak 1`
#> A flowSet with 3 experiments.
#> 
#> An object of class 'AnnotatedDataFrame'
#>   rowNames: s5a01 s5a02 s5a03
#>   varLabels: Patient Visit ... population (6 total)
#>   varMetadata: labelDescription
#> 
#>   column names:
#>   FSC-H SSC-H FL1-H FL2-H FL3-H FL2-A FL4-H Time
#>