Class "filters" and "filtersList"

The filters class is the container for a list of filter objects.

The filtersList class is the container for a list of filters objects.

filters(x)

filtersList(x)

Arguments

x
A list of filter or filters objects.

Value

A filters or filtersList object from the constructor

Details

The filters class mainly exists for displaying multiple filters/gates on one single panel(flowFrame) of xyplot. Note that it is different from filterList class which is to be applied to a flowSet. In other words, filter objects of a fliterList are to be applied to different flowFrames. However,all of filter objects of a filters object are for one single flowFrame, more specifically for one pair of projections(parameters).So these filters should share the common parameters.

And filtersList is a list of filters objects, which are to be applied to a flowSet.

Slots

.Data

Object of class "list". The class directly extends list, and this slot holds the list data.

Extends

Class "list"

Objects from the Class

Objects are created from regular lists using the constructors filters and filtersList:

filters(x)

filtersList(x)

See also