Class "normalization"

Class and methods to normalize a a flowSet using a potentially complex normalization function.

normalization(parameters, normalizationId="defaultNormalization",
              normFunction, arguments=list())

normalize(data, x,...)

Arguments

parameters
Character vector of parameter names.

normalizationId
The identifier for the normalization object.

x
An object of class flowSet.

normFunction
The normalization function

arguments
The list of additional arguments to normFunction

data
The flowSet to normalize.

...
other arguments: see normalize-methodsfor details.

Value

A normalization object for the constructor.

A flowSet for the normalize methods.

Details

Data normalization of a flowSet is a rather fuzzy concept. The idea is to have a rather general function that takes a flowSet and a list of parameter names as input and applies any kind of normalization to the respective data columns. The output of the function has to be a flowSet again. Although we don't formally check for it, the dimensions of the input and of the output set should remain the same. Additional arguments may be passed to the normalization function via the arguments list. Internally we evaluate the function using do.call and one should check its documentation for details.

Currently, the most prominent example for a normalization function is warping, as provided by the flowStats package.

Slots

parameters

Object of class "character". The flow parameters that are supposed to be normalized by the normalization function.

normalizationId

Object of class "character". An identifier for the object.

normFunction

Object of class "function" The normalization function. It has to take two mandatory arguments: x, the flowSet, and parameters, a character of parameter names that are to be normalized by the function. Additional arguments have to be passed in via arguments.

arguments

Object of class "list" A names list of additional arguments. Can be NULL.

Objects from the Class

Objects should be created using the constructor normalization(). See the Usage and Arguments sections for details.

Methods

identifier<-

signature(object = "normalization", value = "character"): Set method for the identifier slot.

identifier

signature(object = "normalization"): Get method for the identifier slot.

normalize

signature(data = "flowSet", x = "normalization"): Apply a normalization to a flowSet.

parameters

signature(object = "normalization"): The more generic constructor.