Define filter boundaries

Convenience methods to facilitate the construction of filter objects

polytopeGate(..., .gate, b, filterId="defaultPolytopeGate")

Arguments

filterId
An optional parameter that sets the filterId of this gate.

.gate
A definition of the gate. This can be either a list, vector or matrix, described below.

b
Need documentation

...
You can also directly describe a gate without wrapping it in a list or matrix, as described below.

Value

Returns a rectangleGate or polygonGate object for use in filtering flowFrames or other flow cytometry objects.

Details

These functions are designed to be useful in both direct and programmatic usage.

For rectangle gate in n dimensions, if n=1 the gate correspond to a range gate. If n=2, the gate is a rectangle gate. To use this function programmatically, you may either construct a list or you may construct a matrix with n columns and 2 rows. The first row corresponds to the minimal value for each parameter while the second row corresponds to the maximal value for each parameter. The names of the parameters are taken from the column names as in the third example.

Rectangle gate objects can also be multiplied together using the * operator, provided that both gate have orthogonal axes.

For polygon gate, the boundaries are specified as vertices in 2 dimensions, for polytope gate objects as vertices in n dimensions.

Polytope gate objects will represent the convex polytope determined by the vertices and parameter b which together specify the polytope as an intersection of half-spaces represented as a system of linear inequalities, \(Ax\le b\)

For quadrant gates, the boundaries are specified as a named list or vector of length two.

See also